quasar/where: move to experimental/ and register SfpuType::where
Align with the abs/fill/trigonometry layout for AI-generated Quasar
SFPU kernels. Adds SfpuType::where so a future ternary-SFPU
dispatcher on Quasar (mirroring Blackhole's
llk_math_eltwise_ternary_sfpu.h) can branch on it without a follow-up
enum edit.

No behavior change: the existing test calls _calculate_where_
directly and is unaffected by the enum entry.
