+++ tt_metal/tt-llk/tests/sources/quasar/sfpu_where_quasar_test.cpp
@@ -77,11 +77,11 @@ const bool is_int_fpu_en = false;
 
 #include "cfg_defines.h"
 #include "cmath_common.h"
+#include "experimental/ckernel_sfpu_where.h"
 #include "llk_math_common.h"
 #include "llk_math_eltwise_unary_datacopy.h"
 #include "llk_math_eltwise_unary_sfpu_common.h"
 #include "params.h"
-#include "sfpu/ckernel_sfpu_where.h"
 
 using namespace ckernel;
 using namespace ckernel::math;
